rft: Delegate responsibility
  - Change TimeSource to abstract base class again
  - Delegate to TimeSourceImplementation that how implement registering observer and notify to observers.
  - This commit's purpose is Child Class of MocktimeSource(ex Clock) doesn't have to know about register observer and notify to them.
