+++ base/time_source.py
@@ -1,13 +1,9 @@
+from abc import ABCMeta, abstractmethod
+
 from base.clock_observer import ClockObserver
 
 
-class TimeSource:
-    def __init__(self):
-        self.__its_observer = []
-
+class TimeSource(metaclass=ABCMeta):
+    @abstractmethod
     def register_observer(self, observer: ClockObserver):
-        self.__its_observer.append(observer)
-
-    def notify(self, hours: int, minutes: int, seconds: int):
-        for observer in self.__its_observer:
-            observer.update(hours, minutes, seconds)
+        pass
